Release version 1.9.2
2015-01-08
In proc SetDefanalLater, the attempt to change to a watch cursor and
update the screen isn't working right because it may change the value
reported by the spin button in Analyze Meals after it has been
verified, so that there is an error pop-up.  Performance is better
since I wrote this part, so I just removed the whole sequence as it
takes only a second or so even with 1200 meals, although there is
some unexplained flickering if there is auto portion control in
effect.  Also, commented out the prepared statement cache set to 100
because it may be that it is slowing down the program if it runs for
several days; however, don't know for sure because it always takes
several days to test.